client-agent/requirements.txt
# Client Agent Dependencies

# Blockchain & Crypto
web3==6.15.0
eth-account==0.11.0

# Environment & Configuration
python-dotenv==1.0.0

# AWS Integration
boto3==1.34.34

# HTTP Client
httpx==0.26.0

# Data Models
pydantic==2.5.3
